==> workflow/scripts/05_peak_annotation_functions.R <==
##########################################################################################
############### unique, all read count input - then merge
##########################################################################################
FtrCount_input<-function(count.file){
  FTR_df=read.delim(count.file, header=T, sep="\t",
                    stringsAsFactors = F,comment.char = '#')
  colnames(FTR_df)[7]='Counts'
  
  #remove rows without start,end
  FTR_df=FTR_df[!is.na(FTR_df$Start) | !is.na(FTR_df$End),]
  
  #create unique ids
  FTR_df$ID=paste0(FTR_df$Chr,":",FTR_df$Start,"-",FTR_df$End,"_",FTR_df$Strand)
  
  return(FTR_df)
}

#########################################################################################
############### splice junctions
##########################################################################################
Join_Junction_original=function(FtrCount,FtrCount_fracJCount) {    
  #merge JCount with unique/all merge
  FtrCount_fracJCount=FtrCount_fracJCount %>%
    merge(
      FtrCount[,c("ID","strand")],
      by.x="PrimaryGene",
      by.y="ID",
      all.x=T) %>%
    separate(
      PrimaryGene,
      into=c('ID','strand'),
      sep = "_",
      remove = F) %>%
    separate(
      ID,
      into=c('chr','start','end'),
      sep = ":|-",
      remove = T)
  print(paste0('check rows: merge jcounts - ',nrow(FtrCount_fracJCount)))
  
  #remove junctions where splicing is within peak
  site_1=FtrCount_fracJCount$Site1_location>=FtrCount_fracJCount$start&
    FtrCount_fracJCount$Site1_location<=FtrCount_fracJCount$end
  site_2=FtrCount_fracJCount$Site2_location>=FtrCount_fracJCount$start&
    FtrCount_fracJCount$Site2_location<=FtrCount_fracJCount$end
  FtrCount_fracJCount=FtrCount_fracJCount[site_1!=site_2,]
  FtrCount_fracJCount$JunctionID=paste0(FtrCount_fracJCount$Site1_chr,':',
                                        FtrCount_fracJCount$Site1_location,'-',
                                        FtrCount_fracJCount$Site2_location,'_',
                                        FtrCount_fracJCount$strand)
  #renumber rownames
  FtrCount_fracJCount$rID = 1:nrow(FtrCount_fracJCount)
  
  #create GR objects
  jcount1.GR=GRanges(seqnames = as.character(FtrCount_fracJCount$Site1_chr), 
                     ranges=IRanges(start = as.numeric(FtrCount_fracJCount$Site1_location), 
                                    end = as.numeric(FtrCount_fracJCount$Site1_location)),
                     strand = FtrCount_fracJCount$strand,
                     PrimaryGene=FtrCount_fracJCount$PrimaryGene,
                     count=FtrCount_fracJCount$counts,
                     JunctionID=FtrCount_fracJCount$JunctionID,
                     rID=FtrCount_fracJCount$rID)
  jcount2.GR=GRanges(seqnames = as.character(FtrCount_fracJCount$Site2_chr), 
                     ranges=IRanges(start = as.numeric(FtrCount_fracJCount$Site2_location), 
                                    end = as.numeric(FtrCount_fracJCount$Site2_location)),
                     strand = FtrCount_fracJCount$strand,
                     PrimaryGene=FtrCount_fracJCount$PrimaryGene,
                     count=FtrCount_fracJCount$counts,
                     JunctionID=FtrCount_fracJCount$JunctionID,
                     rID=FtrCount_fracJCount$rID)
  FtrCount.GR=GRanges(seqnames = as.character(FtrCount$chr), 
                      ranges=IRanges(start = as.numeric(FtrCount$start), 
                                     end = as.numeric(FtrCount$end)),
                      strand = FtrCount$strand,
                      ID=FtrCount$ID)
  
  ############### Add Junction ID
  junctionID <- function(Ftr_in, j_in, j_name, p_name){
    xo=suppressWarnings(as.data.frame(GenomicRanges::findOverlaps(j_in, 
                                                                  Ftr_in, 
                                                                  type = "any",
                                                                  ignore.strand=F)))
    qh=as.data.frame(j_in[xo$queryHits],row.names = NULL)
    colnames(qh)=paste0(colnames(qh),j_name)
    
    sh=as.data.frame(Ftr_in[xo$subjectHits],row.names = NULL)
    colnames(sh)=paste0(colnames(sh),p_name)
    
    return(cbind(qh,sh))
  }
  
  ### Junction 1 ID
  Junc_PL1 = junctionID(FtrCount.GR, jcount1.GR,"_Junction1","_Peaks1")
  ### Junction 2 ID
  Junc_PL2 = junctionID(FtrCount.GR, jcount2.GR,"_Junction2","_Peaks2")
  
  #merge and filter
  Junc_PL=merge(Junc_PL1,Junc_PL2,by.x='rID_Junction1',by.y='rID_Junction2')[,c('PrimaryGene_Junction1',
                                                                                'JunctionID_Junction1',
                                                                                'strand_Peaks1',
                                                                                'ID_Peaks1',
                                                                                'ID_Peaks2')]
  Junc_PL= Junc_PL %>%
    base::subset(!is.na(ID_Peaks1) | !is.na(ID_Peaks2)) %>%
    dplyr::distinct(ID_Peaks1,ID_Peaks2, .keep_all= TRUE)
  
  #create join id col
  Junc_PL$JoinID=paste0(Junc_PL$ID_Peaks1,",",Junc_PL$ID_Peaks2)
  
  ############### Combine all connected peaks
  #create output df
  Junc_PL$matchid=paste(Junc_PL$ID_Peaks1,Junc_PL$ID_Peaks2,sep="|")
  Junc_PL = Junc_PL %>%
    separate(ID_Peaks1,
             sep = "_",
             into = c('ID','strand'),remove = F) %>%
    separate(ID, 
             sep = ":|-",
             into = c('chr','start','end'),
             remove = T)
  
  Junc_PL[c("start","end")] <- sapply(Junc_PL[c("start","end")],as.numeric)
  
  PGene_TBL2 = setNames(data.frame(matrix(ncol = 9, nrow = length(unique(Junc_PL$JoinID))),row.names =unique(Junc_PL$matchid)), 
                        c("PrimaryGene_Junction1", "JunctionID_Junction1", "JoinID","chr",
                          "start","end","strand","ID_Peaks1","ID_Peaks2"))   
  
  Junc_PL_short=Junc_PL
  for (rowid in unique(Junc_PL$matchid)){
    #generate search id
    #IE "chr1:100007068-100007156_\\+|chr1:100082000-100082149_\\+"
    #determine which rows in df partially match any part of search id
    
    match_df=Junc_PL_short[grepl(gsub("_","_\\\\",rowid),
                                 Junc_PL_short$JoinID),]
    if (nrow(match_df)==0) {next}
    
    ## search only peaks in same chrm of interest
    Junc_PL_chrslct=Junc_PL_short[Junc_PL_short$chr%in%unique(match_df$chr),]
    
    #until we find all rows matching all partial matches
    match_original=str_split(rowid,"\\|")%>%unlist()%>%unique()
    match_current=str_split(match_df$matchid,"\\|")%>%unlist()%>%unique()
    
    
    while(length(setdiff(match_current,match_original))>0){ 
      for (rowid2 in setdiff(match_current,match_original)){
        #determine which rows in df partially match any part of search id
        rowid2=paste(rowid2,collapse = '|')
        match_df2=Junc_PL_chrslct[grepl(gsub("_","_\\\\",rowid2),
                                        Junc_PL_chrslct$matchid),]
        
        #original list is current, and current list is new matched df
        match_original=match_current
        match_add=str_split(match_df2$matchid,"\\|")%>%unlist()%>%unique()
        
        match_current=unique(c(match_add,match_current))
        match_df=rbind(match_df,match_df2) %>%
          distinct
      }
    }
    #update df
    PGene_TBL2[rowid,] = list(paste(sort(unique(match_df$PrimaryGene_Junction1)),collapse = ','),
                              paste(sort(unique(match_df$JunctionID_Junction1)),collapse = ','),
                              paste(sort(unique(match_df$JoinID)),collapse = ','),
                              unique(match_df$chr),
                              min((match_df[,c('start','end')])),
                              max((match_df[,c('start','end')])),
                              unique(as.character(match_df$strand_Peaks1)),
                              paste(sort(unique(match_df$ID_Peaks1)),collapse = ','),
                              paste(sort(unique(match_df$ID_Peaks2)),collapse = ','))
    
    #### Remove Joined ID from master junction table Junc_PLOut
    # Junc_PLOut=Junc_PLOut[idcomb$ID_Peaks2%in%Junc_PLOut==F,]
    Junc_PL_short=Junc_PL_short[grepl(gsub("_","_\\\\",paste(unique(c(match_df$ID_Peaks1,match_df$ID_Peaks2)),collapse = "|")),Junc_PL_short$JoinID)==F,]
  }
  
  ### filter
  PGene_TBL2=PGene_TBL2 %>%
    subset(!is.na(PrimaryGene_Junction1)) %>%
    distinct(.keep_all=TRUE)
  
  #create linked id
  PGene_TBL2$linkedID=paste0(PGene_TBL2$chr,":",
                             PGene_TBL2$start,"-",
                             PGene_TBL2$end,"_",
                             PGene_TBL2$strand)
  
  PGene_TBL2=PGene_TBL2[duplicated(PGene_TBL2)==F,]
  
  
  ############### Get all peaks with junctions
  Junc_peaks=unique(c(Junc_PL$ID_Peaks1,Junc_PL$ID_Peaks2))
  
  #Trim peaks without Splicing
  # FtrCount=FtrCount[FtrCount$ID%in%Junc_peaks==F & FtrCount$Counts_fracMM>=read_depth,]
  FtrCount_trimmed=rbind(FtrCount[FtrCount$ID%in%Junc_peaks==F & FtrCount$Counts_fracMM>=read_depth,],## peaks not spliced and count > readDepth cut
                         FtrCount[FtrCount$ID%in%Junc_peaks,]) ## all peaks with spliced reads 
  FtrCount_trimmed$Junc_Peaks=FtrCount_trimmed$ID%in%Junc_peaks
  print(paste0('Junc_Peaks recorded in outputtable: ',
               (length(Junc_peaks)==sum(FtrCount_trimmed$Junc_Peaks))))
  
  out=list(FtrCount_trimmed,PGene_TBL2,Junc_peaks)
  names(out)=c('FtrCount_trimmed','PGene_TBL2','Junc_peaks')
  return(out)
}          

Join_Junction_WrongRead=function(FtrCount,splice_table) {    
  
  # spliceGroups=fread(splice_table, sep="\t",stringsAsFactors = F,data.table=F,fill = T,header = F) 
  ## failed because of some error Expected 20 fields but found 30 even with fill so used read.table (slower but working)
  
  spliceGroups=read.table(splice_table, sep="\t",stringsAsFactors = F,fill = T,header = F) 
  
  rownames(spliceGroups)=paste0('linkedID_',c(1:nrow(spliceGroups)))
  
  spliceGroups_All=spliceGroups
  spliceGroups_All$linkedID=paste0('linkedID_',c(1:nrow(spliceGroups_All)))
  
  spliceGroups_All=reshape2::melt(spliceGroups_All,id.vars='linkedID')
  spliceGroups_All=spliceGroups_All[(spliceGroups_All$value%in%"")==F,]
  
  colapse=function(x){
    # y=  x[x%in%""==F]
    # print(length(y))
    x=  x[x%in%""==F]%>%sort()%>%paste(.,collapse = ',')
    
    return(x)
  }
  spliceGroups=apply(spliceGroups,1,colapse)%>%as.data.frame()
  colnames(spliceGroups)='JoinID'
  spliceGroups=spliceGroups%>%filter(grepl(",",spliceGroups$JoinID))
  
  spliceGroups$linkedID=rownames(spliceGroups)
  
  # trim splice groups with count less then read_depth
  ###################################################
  
  SpliceGroupsCount=function(i,FtrCount,read_depth){
    Trnsc=(i)%>%as.data.frame()
    # print(View(Trnsc));remove(Trnsc);xxxx
    
    rownames(Trnsc)=NULL
    TrnscUL=unlist(unique(str_split(Trnsc$JoinID,pattern = ",")))
    d2=FtrCount[FtrCount$ID%in%TrnscUL,]
    Trnsc[,'Counts_fracMM']=sum(as.numeric(d2[,'Counts_fracMM']))
    
    return(((Trnsc)))
  }
  # system.time({
  spliceGroups=split(spliceGroups, row.names(spliceGroups))%>%mclapply(SpliceGroupsCount ,FtrCount=FtrCount,read_depth=read_depth,mc.cores=8)
  spliceGroups= do.call(rbind, spliceGroups)%>%as.data.frame()
  # })
  
  
  spliceGroups=spliceGroups[spliceGroups$Counts_fracMM>=read_depth,]
  spliceGroups_All=spliceGroups_All[spliceGroups_All$linkedID%in%spliceGroups$linkedID,]
  
  ##Trim peaks without Splicing
  #######################################
  
  
  FtrCount_trimmed=rbind(FtrCount[FtrCount$ID%in%spliceGroups_All$value==F & FtrCount$Counts_fracMM>=read_depth,],## peaks not spliced and count > readDepth cut
                         FtrCount[FtrCount$ID%in%spliceGroups_All$value,]) ## all peaks with spliced reads 
  FtrCount_trimmed$Junc_Peaks=FtrCount_trimmed$ID%in%spliceGroups_All$value
  
  spliceGroups=spliceGroups[rownames(spliceGroups)%in%unique(spliceGroups_All[spliceGroups_All$value%in%FtrCount_trimmed$ID,'linkedID']),]
  
  
  rownames(spliceGroups)=gsub(",","|",spliceGroups$JoinID)
  
  
  # View(FtrCount_out$FtrCount_trimmed )
  # View(FtrCount_out$PGene_TBL2)
  # FtrCount_out$Junc_peaks
  
  out=list(FtrCount_trimmed,spliceGroups,(spliceGroups_All$value))
  names(out)=c('FtrCount_trimmed','PGene_TBL2','Junc_peaks')
  return(out)
}

Join_Junction=function(FtrCount,splice_table) {    
  
  splice_table3=gsub('OneDrive - National Institutes of Health','OneDrive\\\\ -\\\\ National\\\\ Institutes\\\\ of\\\\ Health',splice_table)
  
  cat('
awk ',"'",'{gsub("\\t",",")}1',"' ",splice_table3,' > ',gsub('.txt',"",splice_table3),'2.txt
',sep="",                 
      file = paste0(gsub('.txt',"",splice_table),"_rename.sh"),
      append = F)
  
  system(paste0('bash ',gsub('.txt',"",splice_table3),"_rename.sh"))
  
  
  # spliceGroups=read.table(paste0(gsub('.txt',"",splice_table),'2.txt'), sep="\t",stringsAsFactors = F,fill = T,header = F) 
  spliceGroups=fread(paste0(gsub('.txt',"",splice_table),'2.txt'), sep="\t",stringsAsFactors = F,data.table=F,fill = T,header = F)
  ##########    
  spliceGroups=spliceGroups%>%filter(grepl(",",spliceGroups$V1))
  
  rownames(spliceGroups)=paste0('linkedID_',c(1:nrow(spliceGroups)))
  
  
  
  spliceGroups_All=apply(spliceGroups,1,str_split,pattern=",")
  spliceGroups_All=reshape2::melt(spliceGroups_All)
  spliceGroups_All=dplyr::rename(spliceGroups_All,'linkedID'=L1)
  
  colnames(spliceGroups)='JoinID'
  spliceGroups$linkedID=rownames(spliceGroups)
  
  
  # trim splice groups with count less then read_depth
  ###################################################
  
  SpliceGroupsCount=function(i,FtrCount,read_depth){
    Trnsc=(i)%>%as.data.frame()
    # print(View(Trnsc));remove(Trnsc);xxxx
    
    rownames(Trnsc)=NULL
    TrnscUL=unlist(unique(str_split(Trnsc$JoinID,pattern = ",")))
    d2=FtrCount[FtrCount$ID%in%TrnscUL,]
    Trnsc[,'Counts_fracMM']=sum(as.numeric(d2[,'Counts_fracMM']))
    
    return(((Trnsc)))
  }
  # system.time({
  spliceGroups=split(spliceGroups, row.names(spliceGroups))%>%mclapply(SpliceGroupsCount ,FtrCount=FtrCount,read_depth=read_depth,mc.cores=8)
  spliceGroups= do.call(rbind, spliceGroups)%>%as.data.frame()
  # })
  
  
  spliceGroups=spliceGroups[spliceGroups$Counts_fracMM>=read_depth,]
  spliceGroups_All=spliceGroups_All[spliceGroups_All$linkedID%in%spliceGroups$linkedID,]
  
  ##Trim peaks without Splicing
  #######################################
  
  
  FtrCount_trimmed=rbind(FtrCount[FtrCount$ID%in%spliceGroups_All$value==F & FtrCount$Counts_fracMM>=read_depth,],## peaks not spliced and count > readDepth cut
                         FtrCount[FtrCount$ID%in%spliceGroups_All$value,]) ## all peaks with spliced reads 
  FtrCount_trimmed$Junc_Peaks=FtrCount_trimmed$ID%in%spliceGroups_All$value
  
  spliceGroups=spliceGroups[rownames(spliceGroups)%in%unique(spliceGroups_All[spliceGroups_All$value%in%FtrCount_trimmed$ID,'linkedID']),]
  
  
  rownames(spliceGroups)=gsub(",","|",spliceGroups$JoinID)
  
  
  # View(FtrCount_out$FtrCount_trimmed )
  # View(FtrCount_out$PGene_TBL2)
  # FtrCount_out$Junc_peaks
  
  out=list(FtrCount_trimmed,spliceGroups,(spliceGroups_All$value))
  names(out)=c('FtrCount_trimmed','PGene_TBL2','Junc_peaks')
  return(out)
}

##########################################################################################
############### DEP
##########################################################################################
DEP_input=function(FtrCount_trimmed){
  print(paste0("Running ",DEmethod," Peaks file prep"))
  
  manorm_bed=FtrCount_trimmed[,c('chr','start','end','ID','ID','strand')]
  manorm_bed[,5]=1
  
  #write peaks bed file
  write.table(manorm_bed,
              file=paste0(out_dir_DEP, sample_id, "_", peak_type ,"readPeaks_for",DEmethod,".bed"), 
              sep = "\t", row.names = FALSE, col.names = F, append = F, quote= FALSE,na = "")
  #write peaks + bed file
  write.table(manorm_bed[manorm_bed$strand%in%"+",],
              file=paste0(out_dir_DEP, sample_id, "_", peak_type ,"readPeaks_for",DEmethod,"_P.bed"), 
              sep = "\t", row.names = FALSE, col.names = F, append = F, quote= FALSE,na = "")
  
  #write peaks - bed file
  write.table(manorm_bed[manorm_bed$strand%in%"-",],
              file=paste0(out_dir_DEP, sample_id, "_", peak_type ,"readPeaks_for",DEmethod,"_N.bed"), 
              sep = "\t", row.names = FALSE, col.names = F, append = F, quote= FALSE,na = "")
}

##########################################################################################
############### reference, annotation table
##########################################################################################
## need to optomize because reading in tables twice when using different Output_option
## Will not take longer with larger datasets
#gencode or Combined. option refers to which reftable to output. Gencode or rpmk+Custom+refseq
Ref_Prep=function(Output_option){
  print("Prep References")
  
  #read in references
  ref_table = read.table(reftable_path, header = TRUE, sep="\t", row.names = 1)
  
  gencodeCol=c('chr','start','end','strand','ensembl_gene_id',
               'transcript_id','external_gene_name','gene_type','gene_type_ALL')
  ref_gencode_trans = read.table(gencode_transc_path,header=TRUE,sep="\t")
  gencode_Anno_RNA_comb=ref_gencode_trans[,gencodeCol]  
  
  #create annotation dfs
  rpmk_Anno_RNA_comb=data.frame()
  custom_Anno_RNA_comb=data.frame()
  refseq_Anno_RNA_comb=data.frame()
  
  for (rowid in rownames(ref_table)){
    ref_type = ref_table[rowid,'AnnoType']
    ref_selection =  ref_table[rowid,paste0(ref_species,"_selection")]
    ref_selection=unlist(strsplit(ref_selection,split ="," ))  
    rowid=gsub(" ","",rowid)
    
    if (ref_type=='transcript_anno') {
      for (selec in ref_selection) {
        if (selec =='Gencode') {
          Gtable=read.table(paste0(anno_dir,rowid,'_',selec,'.bed'),header=T,sep="\t")
          gencode_Anno_RNA_comb=rbind(gencode_Anno_RNA_comb,Gtable[,gencodeCol])
        } else if (selec =='Repeatmasker') {
          Rtable=read.table(paste0(anno_dir,rowid,'_',selec,'.bed'),header=TRUE,sep="\t")
          rpmk_Anno_RNA_comb=rbind(rpmk_Anno_RNA_comb,Rtable)
        } else if (selec =='Custom') {
          Ctable=read.table(paste0(anno_dir,rowid,'_',selec,'.bed'),header=TRUE,sep="\t")
          custom_Anno_RNA_comb=rbind(custom_Anno_RNA_comb,Ctable)
        } else if (selec =='refseq') {
          Reftable=read.table(paste0(anno_dir,rowid,'_',selec,'.bed'),header=TRUE,sep="\t")
          refseq_Anno_RNA_comb=rbind(refseq_Anno_RNA_comb,Reftable)
        }
        
      }
    }else { next }
  }  
  
  gencode_Anno_RNA_comb=gencode_Anno_RNA_comb[duplicated(gencode_Anno_RNA_comb)==F,]
  
  #prepare annotations
  prep_annotation<-function(df.in,col.in){
    
    df_clean = df.in %>%
      distinct()
    
    if (dim(df_clean)[1] == 0) {
      df_out=as.data.frame(matrix(ncol=length(c("chr","start","end","strand","type","transcript_name")),nrow=0))
      colnames(df_out)=c("chr","start","end","strand","type","transcript_name")
    } else {
      df_out=df_clean[,col.in]
      colnames(df_out)=c("chr","start","end","strand","type","transcript_name")
    }
    return(df_out)
  }
  
  Anno_RNA_comb = rbind(
    prep_annotation(rpmk_Anno_RNA_comb,c("chr","start","end","strand","type","name")),
    prep_annotation(custom_Anno_RNA_comb,c("chr","start","end","strand","type","name")),
    prep_annotation(refseq_Anno_RNA_comb,c("chr","start","end","strand","gene_type","ensembl_gene_id"))) %>%
    distinct()
  Anno_RNA_comb=Anno_RNA_comb[is.na(Anno_RNA_comb$chr)==F,]
  
  if(Output_option=='gencode'){return(gencode_Anno_RNA_comb)}
  else if(Output_option=='Combined'){return(Anno_RNA_comb)}
  else{print('slection error')}
  
}

##########################################################################################
############### call peaks
###########################################################################################
#merge references depending on species, run annotation
bam_anno2<-function(peaksTable,Annotable,ColumnName,pass_n){
    # peaksTable=peak_in
    # Annotable=gencode_Anno_RNA_comb[,colMerge]
    # ColumnName=colSelect
    # pass_n="pass1"

  #set file prefix
  file_prefix = paste0(file_id,pass_n,"_")

##########################
## Set up tables for bedtools intersect
##########################  
  
  ## the multiple chr is a space filler was to make a bed6 file for bedtools 
  anno_output = Annotable[,c('chr','start','end','chr','chr','strand',ColumnName)] %>%
    dplyr::rename(
      chr_anno = chr,
      start_anno = start,
      end_anno = end,
      strand_anno = strand
    )
  
  #clear chr cols 4,5
  anno_output[,4:5]="."
  
  #output bed files
  write.table(anno_output,
              file = paste0(tmp_dir,file_prefix,"annotable.bed"), 
              sep = "\t", row.names = FALSE, col.names = F, append = F, quote= FALSE)
  
  #write out peakstable
  peaksTable_output=subset(peaksTable,!is.na(peaksTable$start) & 
                             chr %in% unique(anno_output$chr_anno))[,c('chr','start','end','ID','ID','strand')]
  write.table(peaksTable_output,
              file = paste0(tmp_dir,file_prefix,"peakstable.bed"), 
              sep = "\t", row.names = FALSE, col.names = F, append = F, quote= FALSE)
  
  #merge bedfiles into output text
  system(paste0('bedtools intersect -a ',
                tmp_dir, file_prefix,'peakstable.bed -b ',
                tmp_dir, file_prefix,'annotable.bed -wao -s > ',
                tmp_dir, file_prefix,'peaks_OLab.txt'))
  
  #read in merged file
  ab_OL = read.table(paste0(tmp_dir,file_prefix,"peaks_OLab.txt"), 
                     header=F, sep="\t",stringsAsFactors = F)
  
  file.remove(paste0(tmp_dir,file_prefix,"peaks_OLab.txt"))
  
  colnames(ab_OL) = c(paste0(colnames(peaksTable_output)),
                      paste0(colnames(anno_output)),'ntOL')

  ##################################################################################################################################    
##########################
## format annotatin intersect table
##########################  
  #add width
  ab_OL$width = ab_OL$end-ab_OL$start
  ab_OL$width_anno = ab_OL$end_anno-ab_OL$start_anno
  
  #subset for annotations with annotations
  ab_OL=ab_OL[ab_OL$ntOL>0,]
  
  if (nrow(ab_OL)>0){
    #add nt / width in annotation and read
    ab_OL$OLper_anno = ab_OL$ntOL/ab_OL$width_anno ## % overlap region+annotation by anno size
    ab_OL$OLper = ab_OL$ntOL/ab_OL$width  ## % overlap region+annotation by peak size
    
    #subset for annotations where overlap with peak is >50% 
    #   # or for large peaks get annotations when peak covers > 75% of annotation
    ab_OL=ab_OL[(ab_OL$OLper_anno>.75 | ab_OL$OLper>.51), ]
    
    #ID duplicates and unique
    dup=unique(ab_OL[duplicated(ab_OL$ID),'ID']) ### regions with multiple annotations
    ab_OL_single=ab_OL[!(ab_OL$ID%in%dup),]  ### table of regions with one annotation
    ab_OL_double=ab_OL[(ab_OL$ID%in%dup),]  ### table of all annotations for regions with multiple annotations

    ####################################        
    # Process table of all annotations for regions with multiple annotations 
    ####################################
    unique_list=unique(ab_OL_double$ID)
    ab_OL_colapsed=as.data.frame(matrix(nrow=length(unique_list),ncol=ncol(ab_OL_double)));
    colnames(ab_OL_colapsed)=colnames(ab_OL_double)
    
    #for every unique id with multiple annotations
    for(x in 1:length(unique_list)){
      peaksTableU = unique_list[x]
      pam = ab_OL_double[ab_OL_double$ID%in%peaksTableU,]
        ##get all anno info from first anno of region
      ab_OL_colapsed[x,colnames(ab_OL_double)%in%c(ColumnName)==F]=(pam[1,colnames(ab_OL_double)%in%c(ColumnName)==F,drop=T])
        ##Collapse desired columns
      for (cx in 1:length(ColumnName)) {
        ab_OL_colapsed[x,ColumnName[cx]]=paste(sort(unique((pam[,ColumnName[cx]]))), collapse =",")
      }
    }
    
    ### Combine regions with single and multiplle annoations (region should only occure once in table)
    ab_OL_colapsed=rbind(ab_OL_colapsed,ab_OL_single)
    ab_OL_colapsed=ab_OL_colapsed[!is.na(ab_OL_colapsed$ID),]
    ab_OL_colapsed[(ab_OL_colapsed$ntOL<=0),ColumnName]=NA
    
    ###fix rename variable
    print(paste0('Overlaps with ',nrow(ab_OL_colapsed),' Regions from '))[1]
    
    #merge tables
    ab_OL_colapsed = merge(peaksTable,
                           ab_OL_colapsed[,c('ID',ColumnName)],
                           by='ID',all.x=T)
    
  } else{ 
    print("does not overlap with ")[1]
    # next
    ab_OL_colapsed=peaksTable
    ab_OL_colapsed[,ColumnName]=NA
  }
  
  print((match.call())$Annotable)
  
  return(ab_OL_colapsed) 
}

peak_calling<-function(peak_in,nmeprfix){
  # peak_in=FtrCount_trimmed[,c('chr','start','end','ID','ID','strand')] #peaks_Oppo
  # nmeprfix="Same_"
  
  colMerge = c('chr','start','end','strand','ensembl_gene_id','transcript_id',
               'external_gene_name','gene_type','gene_type_ALL')
  colSelect=c('ensembl_gene_id','external_gene_name','gene_type','gene_type_ALL')

  system.time({ print('bam_anno2')
    
  PeaksdataOut=bam_anno2(peak_in,
                         gencode_Anno_RNA_comb[,colMerge],
                         colSelect,
                         "pass1")
  print("D1")
  PeaksdataOut=bam_anno2(PeaksdataOut,
                         Anno_RNA_comb,
                         c('type','transcript_name'),
                         "pass2")
  print("D2")
  })
  ##########################################################################################
  ############### Clean up
  ##########################################################################################
  # change lincRNA,rRNA to rRNA only
  # do not change the name if there is actually a lincRNA name separate from the rRNA grepl(',',p$name)==F)
  PeaksdataOut[(grepl(',',PeaksdataOut$name)==F)&(PeaksdataOut$type%in%'lincRNA-exon,rRNA'),'type']='rRNA'
  PeaksdataOut[(grepl(',',PeaksdataOut$name)==F)&(PeaksdataOut$type%in%'lincRNA-intron,rRNA'),'type']='rRNA'
  PeaksdataOut[(grepl(',',PeaksdataOut$name)==F)&(PeaksdataOut$type%in%'lncRNA,rRNA'),'type']='rRNA'
  
  # change all psueudogene classes to pseudogene
  PeaksdataOut[grep('pseudogene',PeaksdataOut$gene_type_ALL),'gene_type_ALL']='pseudogene'
  
  # change all sc RNA to yRNA names
  PeaksdataOut$type=gsub('scRNA,yRNA','yRNA',PeaksdataOut$type)
  
  ##########################################################################################
  ############### Summarize Peak annotation
  ##########################################################################################
  #separate annotated and non annotated peaks 
  anno=PeaksdataOut[(is.na(PeaksdataOut$gene_type_ALL)==F)|(is.na(PeaksdataOut$type)==F),'ID']
  
  #Only subset of peaks are annotated (most likely)
  if (length(anno)<nrow(PeaksdataOut)&length(anno)>0) { 
    peaksTable_Anno=PeaksdataOut[(PeaksdataOut$ID%in%anno),,drop=F]
    peaksTable_noAnno=PeaksdataOut[!(PeaksdataOut$ID%in%anno),,drop=F]
  } else if (length(anno)==nrow(PeaksdataOut)) { #"All peaks are annotated"
    peaksTable_Anno=PeaksdataOut[(PeaksdataOut$ID%in%anno),,drop=F]
    peaksTable_noAnno=as.data.frame(matrix(nrow=1,ncol=ncol(PeaksdataOut)))
    colnames(peaksTable_noAnno)=colnames(PeaksdataOut)
  }
  system.time({ 
  if(length(anno)>0) { 
    
    u=unique(peaksTable_Anno$ID)
    peaksTable_colapsed=as.data.frame(matrix(nrow=length(u),ncol=ncol(peaksTable_Anno)));
    colnames(peaksTable_colapsed)=colnames(peaksTable_Anno)
    
    ## Checking for conflicting annotations from gencode and additional supplied annotations
    ## If conflict prioritize Supplied annotations 
    ## Additionally Peak may overlap multiple features and so I am setting hierarchy

      
    print(paste0('Annotated Regions: ', length(u)))
    for(x in 1:length(u)){
      PeaksdataOutU=u[x]
      pam=peaksTable_Anno[peaksTable_Anno$ID%in%PeaksdataOutU,]
      
      #Comb Annotation
      pam_1=as.character(pam$gene_type_ALL) ### type from Gencode 
      pam_1=as.data.frame(strsplit(pam_1,','));colnames(pam_1)='a';pam_1$a=as.character(pam_1$a)
      
      pam_2=as.character(pam$type) ### type from Annotation
      pam_2=as.data.frame(strsplit(pam_2,','));colnames(pam_2)='a';pam_2$a=as.character(pam_2$a)
# if ((is.na(pam$ensembl_gene_id)==F)&(is.na(pam$transcript_name)==F)) {
#   if (nrow(pam_1)>1) {
#     if (grepl("lnc",pam_1)%in%F) {
#       if (pam_1!=pam_2) {
#         pamtest=pam
#         pam_1test=pam_1
#         pam_2test=pam_2
#         STOP
#       }
#     }
#   }
# }
      type_list = c("miRNA","piRNA","rRNA","siRNA","snRNA","snoRNA",
                    "ribozyme","scRNA", "sRNA","scaRNA", "yRNA", "srpRNA", "7SKRNA",
                    "sncRNA","vaultRNA", "tRNA")
####################
### Format Gencode annotations
######################      
if(nrow(pam_1)>=1&(NA%in%pam_1)==F) {
      #remove misc_RNA catagory : these are covered by additional annotations (mostly yRNA)
      ## We should set this up to only display misc_RNA if no additional annotations
      if (nrow(pam_1)>1&grepl('misc_RNA',pam_1)&grepl(paste(type_list,collapse = "|"),pam_1)){
        pam_1[pam_1$a%in%'misc_RNA','a']=NA
      }
      if (grepl('misc_RNA',pam_1)&'FALSE'%in%is.na(pam_2)){
        pam_1[pam_1$a%in%'misc_RNA','a']=NA
      }
      
      #change ribozyme (gencode) to RNA type from additional anno
      if (grepl('ribozyme',pam_1)) {
        pam_1[pam_1$a%in%'ribozyme',]=unique(pam_2$a)
      }

  }
      #lincRNA are from an older Gencode version (VM18 or older) so don't use    
      if (grepl('lncRNA',pam_1)&grepl('lincRNA',pam_2)) {xxx
        pam_1[pam_2$a%in%'lincRNA',]='lncRNA'
      }
      
######################      
## combine all rna types subtypes into ncRNA
######################      
#### Annotation from Gencode : unique(sort(mm10$gene_type_ALL))

ReplaceType <-function(df_in,col_in,list_in,replace_id){
  for (id in list_in){
    df_in[,col_in]=gsub(paste0("\\<",id,"\\>"),replace_id,df_in[,col_in])
  } 
  return(df_in)
}
      type_list = c(type_list,"miscRNA","misc_RNA")

      pam_c=rbind(pam_1,pam_2)
      pam_c=pam_c[!is.na(pam_c$a),,drop=F]
      
      pam_c2 = ReplaceType(pam_c,"a", type_list,"ncRNA")
      
      if (length(pam_c$a)>1) {pam_c=pam_c[order(pam_c$a),,drop=F]}
      if (length(pam_c2$a)>1) {pam_c2=pam_c2[order(pam_c2$a),,drop=F]}
      
      pam_c=paste(unique(pam_c$a),collapse = ',')
      pam_c2=paste(unique(pam_c2$a),collapse = ',')

######################      
## combine GeneName from Gencode transcript_name from additional anno
###################### 
      
      #Comb GeneName 
      pam_G1=as.character(pam$external_gene_name) ### gene name from gencode
      pam_G1=as.data.frame(strsplit(pam_G1,','));colnames(pam_G1)='a';pam_G1$a=as.character(pam_G1$a)
      
      ### G1 is gencode annotation and G2 is additional annotation
      pam_G2=as.character(pam$transcript_name) ### gene name from annotation
      pam_G2=as.data.frame(strsplit(pam_G2,','));colnames(pam_G2)='a';pam_G2$a=as.character(pam_G2$a)
      
      ## there are non uniform naming for 7SK rRNA so i am making a uniform naming by selecting only from gencode
      if (grepl('sk',pam_G1)&grepl('7SK',pam_G2)) {
        pam_G2[pam_G2$a%in%'7SK',]=NA
      }

######################      
## merge gene names from "gencode" and "annotation"
###################### 
      pam_gmerge=rbind(pam_G1,pam_G2)
      pam_gmerge=pam_gmerge[!is.na(pam_gmerge$a),,drop=F]
      pam_gmerge=paste(unique(pam_gmerge$a),collapse = ',')
      
      peaksTable_colapsed[x,colnames(peaksTable_Anno)]=(pam[1,colnames(peaksTable_Anno),drop=T])
      peaksTable_colapsed[x,'type_simple_comb']=pam_c2
      peaksTable_colapsed[x,'type_comb']=pam_c
      peaksTable_colapsed[x,'gene_name_comb']=pam_gmerge
    }
####################################################################################################################################    
    
######################      
## Format output table
######################    
    
    
    ######################      
    ## Comnbine anno regions and NoAnno Regions
    ######################     
    rnatype=peaksTable_colapsed[,c('ID','ensembl_gene_id','external_gene_name','gene_type',"gene_type_ALL",
                                   'type','transcript_name','type_simple_comb','type_comb','gene_name_comb')]
    
    peaksTable_noAnno[,c('type_simple_comb','type_comb','gene_name_comb')]=NA
    
    peaksTable_colapsed=rbind(peaksTable_colapsed,peaksTable_noAnno)
    peaksTable_colapsed=peaksTable_colapsed[!is.na(peaksTable_colapsed$ID),]
    
    rnames=c("ensembl_gene_id","external_gene_name","gene_type","gene_type_ALL","type","transcript_name","type_simple_comb","type_comb","gene_name_comb")
    
    colnames(peaksTable_colapsed)[colnames(peaksTable_colapsed)%in%rnames]=paste0(nmeprfix,rnames)

    ######################      
    ## Rename Lnc Annotations in select columns(col_list)
    ######################       
    type_start = c('lincRNA-exon,lncRNA','lincRNA-intron,lncRNA')
    type_end = c('lincRNA-exon','lincRNA-intron')
    col_list= c('type_comb','type_simple_comb','gene_type_ALL')
    
    #change the type (type_list) to new id (change_list)
    for (cols_to_change in col_list){
      
      #change the type (type_list) to new id (change_list)
      for (i in range(1:length(type_start))){
        peaksTable_colapsed[,paste0(nmeprfix,cols_to_change)] = gsub(type_start[i],
                                                                     type_end[i],
                                                                     peaksTable_colapsed[,paste0(nmeprfix,cols_to_change)] )
      }
    }
####################################################################################################################################    
  } else { 
    print("No Peaks Annotated")
    peaksTable_colapsed=PeaksdataOut
    rnames=c("ensembl_gene_id","external_gene_name","gene_type","gene_type_ALL","type","transcript_name","type_simple_comb","type_comb","gene_name_comb")
    peaksTable_colapsed[,rnames]=NA
    colnames(peaksTable_colapsed)[colnames(peaksTable_colapsed)%in%rnames]=paste0(nmeprfix,rnames)
  }  
  }) 
  return(peaksTable_colapsed)
}

##########################################################################################
############### INTRON EXON ANNOTATION
##########################################################################################
### Identify if CLIP peak overlaps with Intron or Exonic region   

#   Using GTF file from GENCODE v23 -mm10
#   Using GTF file from GENCODE v32 -hg38

# Peaks were annotated by whether they overlap with Host gene intron/exon region
# Intron coordinates were calculated from GTF file.
# A second column was added to idenify if the peak also overlapped with the 5'UTR 3'UTR or CDS (Column: Featrue 2)
#########################################################################################3
IntronExon_prep=function(gencode_path,intron_path,gencode_transc_path){
  #gencode
  ref_gencode = fread(gencode_path, header=T, sep="\t",stringsAsFactors = F,data.table=F) %>%
    subset(transcript_type != "TEC" & gene_type != "TEC") %>% 
    dplyr::rename(
      chr = seqname,
      ensembl_gene_id = gene_id,
      external_gene_name = gene_name
    )
  
  ref_gencode_trans = read.table(gencode_transc_path,header=TRUE,sep="\t")
  
  #exons
  ref_gencode_e = base::subset( ref_gencode, feature == "exon")
  
  #introns
  introns=fread(intron_path, 
                header=F, sep="\t",stringsAsFactors = F,data.table=F, 
                col.names = c('chr','start','end','attribute','V5','strand'))# %>%
    introns=suppressWarnings(separate(introns,attribute,
             into=c('transcript_id','feature','exon_number','level','chr2','intronnumber','dir'),
             remove = T,sep = "_"))
  
  #remove Non-Chromosome contigs
  introns$chr=(gsub('chr[0-9]+_|chr[X-Y]_|chrUn_|_alt|_random|_fix','',introns$chr))
  introns$chr=(gsub('v','.',introns$chr))
  
  #split out transcript id
  introns$transcript_id = unlist(lapply(stringr::str_split(introns$transcript_id, "[.]"), "[[",1))
  
  # intron table is 0 based while exon starts at 1
  # Same for exon number the exon # starts at 0
  introns$start=introns$start+1
  introns$exon_number=as.numeric(introns$exon_number)+1
  
  # add gene type type
  introns=merge(introns,
                unique(ref_gencode_trans[,c('transcript_id',
                                            'ensembl_gene_id','external_gene_name',
                                            'gene_type','gene_type_ALL','transcript_type',
                                            'transcript_name','score')]),
                by='transcript_id',all.x=T)
  
  #binding exon, intron
  intron_exon=rbind(ref_gencode_e[,c('chr','feature','start','end','strand','transcript_id',
                                     'ensembl_gene_id','external_gene_name','gene_type',
                                     'transcript_type','transcript_name','score','exon_number')],
                    introns[,c('chr','feature','start','end','strand','transcript_id',
                               'ensembl_gene_id','external_gene_name','gene_type','transcript_type',
                               'transcript_name','score','exon_number')])
  intron_exon$ID=paste0(intron_exon$chr,':',intron_exon$start,'-',intron_exon$end,"_",intron_exon$strand)
  
  return(intron_exon)
  
}  

##########################################################################################
############### Intron Exon Calling
##########################################################################################
IE_calling <- function(peak_in,Peak_Strand_ref,nmeprfix){
  # peak_in=FtrCount_trimmed[,c('chr','start','end','ID','ID','strand')]
  # Peak_Strand_ref='Same'
  # nmeprfix="Same_"
  
  #create annotation table
  ColumnName = c("feature","exon_number")
  anno_IntExn = intron_exon[grep('protein_coding',intron_exon$gene_type),][,c('chr','start',
                                                                              'end','transcript_id',
                                                                              'transcript_id','strand',ColumnName)]
  
  ###fix variable swap
  p=peak_in[,c('chr','start','end','ID','ID','strand')]
  
  # Setting up bed 6 file but need column to be clear for bedtools intersect
  anno_IntExn[,5]=""
  
  #add _anno to col names
  colnames(anno_IntExn)[colnames(anno_IntExn)%in%c('chr','start','end','strand')]=paste0(c('chr','start','end','strand'),'_anno')
  
  ## setting up bedtools intersect
  write.table(anno_IntExn,file=paste0(out_dir,file_id,"annotable.bed"), sep = "\t", 
              row.names = FALSE, col.names = F, append = F, quote= FALSE)
  write.table(p,file=paste0(tmp_dir,file_id,"peakstable.bed"), sep = "\t", 
              row.names = FALSE, col.names = F, append = F, quote= FALSE)
  system(paste0('bedtools intersect -a ', tmp_dir, file_id, 'peakstable.bed -b ',
                out_dir, file_id, 'annotable.bed -wao -s  >', tmp_dir, file_id, 'peaks_OLexon.txt'))
  
  exoninof=fread(paste0(tmp_dir,file_id,"peaks_OLexon.txt"), 
                 header=F, sep="\t",stringsAsFactors = F,data.table=F)
  
  file.remove(paste0(tmp_dir,file_id,"peaks_OLexon.txt"))
  
  #set up df
  colnames(exoninof)=c(paste0(colnames(p)),paste0(colnames(anno_IntExn)),'ntOL')
  exoninof=exoninof[exoninof$ntOL>0,]
  exoninof$width_anno=exoninof$end_anno-exoninof$start_anno
  exoninof$OLper_anno=exoninof$ntOL/exoninof$width_anno
  exoninof$width=exoninof$end-exoninof$start
  exoninof$OLper=exoninof$ntOL/exoninof$width
  exoninof=exoninof[(exoninof$OLper_anno>.75 | exoninof$OLper>.51), ]
  
print('Annotate Peaks with IE')

# Find list of IDs 
IDs <- peak_in$ID

# Create empty dataframe
peak_in_df <- data.frame(matrix(ncol = 15, nrow = length(IDs)))
peak_in_df[is.na(peak_in_df)] <- 'NA'

rownames(peak_in_df) <- IDs
colnames(peak_in_df) <- c('chr','start','end','ID','ID.1','strand',
                          paste0(nmeprfix,'Exn_start_dist'),
                          paste0(nmeprfix,'Intron_start_dist'),
                          paste0(nmeprfix,'Intron_5pStart'),
                          paste0(nmeprfix,'Exn_5pStart'),
                          paste0(nmeprfix,'feature'),
                          paste0(nmeprfix,'exon_number'),
                          paste0(nmeprfix,'intron_number'),
                          paste0(nmeprfix,'intron_LargeOL'),
                          paste0(nmeprfix,'exon_LargeOL'))

peak_in_df$chr <- peak_in$chr
peak_in_df$start <- peak_in$start
peak_in_df$end <- peak_in$end
peak_in_df$strand <- peak_in$strand
peak_in_df$ID <- IDs
peak_in_df$ID.1 <- IDs

# Report Number of Peaks
print(paste0('Annotated regions: ', nrow(peak_in)))

# Create the function that will fill up the dataframe
update_rows <- function(peak_in) {

  # Find ID for row
  ID = peak_in['ID']

  # Find all rows with ID in intersect file
  sub_exoninof <- exoninof[which(exoninof$ID == ID),]

  # Make list of unique features
  peak_in[paste0(nmeprfix,'feature')] <- paste0(unique(sub_exoninof$feature), collapse = ',')

  # Separate following computations for exons and introns
  for (feature in c('exon','intron')){

  # Make sub df for exons or introns
  feature_df <- sub_exoninof[which(sub_exoninof$feature == feature),]


  # Separate inton + and - strand for distance computations
  feature_pos_df <- feature_df[which(feature_df$strand == '+'),] # FEATURES ON STRAND +
  feature_neg_df <- feature_df[which(feature_df$strand == '-'),] # FEATURES ON STRAND -

  # Compute distance 
  #Same_Feat_start_dist <- min(c(abs(feature_pos_df$start - feature_pos_df$start_anno) / (feature_pos_df$end_anno - feature_pos_df$start_anno) *100,
   #                             abs(feature_neg_df$end - feature_neg_df$end_anno) / (feature_neg_df$end_anno - feature_neg_df$start_anno) *100))

  # Fill dataframe separately for exons and introns
  # Distance, unique list of exon/intron number, maximum overlap percentage 
  if (feature == 'exon') {

  #peak_in[paste0(nmeprfix,'Exn_start_dist')] <- Same_Feat_start_dist
  peak_in[paste0(nmeprfix,'exon_number')] <- paste0(unique(feature_df$exon_number), collapse = ',')
  peak_in[paste0(nmeprfix,'exon_LargeOL')] <- max(feature_df$OLper)

      } else if (feature == 'intron'){
  #peak_in[paste0(nmeprfix,'Intron_start_dist')] <- Same_Feat_start_dist
  peak_in[paste0(nmeprfix,'intron_number')] <- paste0(unique(feature_df$exon_number), collapse = ',')
  peak_in[paste0(nmeprfix,'intron_LargeOL')] <- max(feature_df$OLper)

      }
    }
  return(peak_in)
  }

# Run function on each row/ID 
peak_in_df <- t(apply(peak_in_df, 1, update_rows)) # Run function on each row / ID

# Substitute Inf and Empty cells
peak_in_df[peak_in_df == -Inf] <- 'NA'
peak_in_df[peak_in_df == Inf] <- 'NA'
peak_in_df[peak_in_df == ''] <- 'NA'

return(peak_in_df)
}

  ### protein coding peaks with not Correctly assigned exon overlap -- Moved to 09_Anno_Process.R

##########################################################################################
############### IDENTIFY PEAKS IN REPEAT REGIONS   
##########################################################################################
# Annotate all repeat regions/Classes identified in Repeatmasker Annotation file (UCSC Table browser)  
# Data was not filtered based on any of the identified Repeats.  
#  1) LINE/SINE   
#  2) LTR   
#  3) DNA   
#  4) Satalites   
#  5) Simple Repeats   
#  6) Low Complexity   
#  7) Other   
#  8) Unknown 

#read in rmsk
rmsk_prep=function(rmsk_path){
  rmsk_GRCm38=fread(rmsk_path, header=T, sep="\t",stringsAsFactors = F,
                    data.table=F)
  rmsk_GRCm38$genoName=gsub('chr[0-9]+_|chr[X-Y]_|chrUn_|_alt|_random|_fix',"",rmsk_GRCm38$genoName)
  rmsk_GRCm38$genoName=gsub("v",".",rmsk_GRCm38$genoName)
  
  return(rmsk_GRCm38)
}

rpmsk_anno <- function(ColumnName,Annotable,peaksTable){
  
  if(length(args)==0){
    nmeprfix="Same_"
    peaksTable=intronexon_Same
    Annotable=rmsk_GRCm38[rmsk_GRCm38$repClass%in%c('LINE','SINE'),]
    ColumnName=paste0(nmeprfix,'Repeat_LINE_SINE')
  }
  
  #create df of ids and start location
  peaksTable=separate(peaksTable,ID,into=c('start','strand'),sep="_",remove=F)
  peaksTable=separate(peaksTable,start,into=c('chr','start','end'),sep=":|-",remove=F)
  
  #create granges obj for peaks
  peaksTable.GR <- GRanges(seqnames = as.character(peaksTable$chr), 
                           ranges=IRanges(start = as.numeric(peaksTable$start),
                                          end = as.numeric(peaksTable$end)),
                           strand = peaksTable$strand,ID=peaksTable$ID )
  
  #create granges obj for annotation
  anno.GR <- GRanges(seqnames = as.character(Annotable$genoName), 
                     ranges=IRanges(start = as.numeric(Annotable$genoStart), 
                                    end = as.numeric(Annotable$genoEnd)),
                     strand = Annotable$strand,repClass=Annotable$repClass,repName=Annotable$repName )
  
  ###fix variable renames
  q =peaksTable.GR
  s=anno.GR
  
  #find annotation overlaps
  xo=suppressWarnings(as.data.frame(GenomicRanges::findOverlaps(q,s,type = "any",ignore.strand=F)))
  
  #run query
  qh=as.data.frame(q[xo$queryHits],row.names = NULL)
  sh=as.data.frame(s[xo$subjectHits],row.names = NULL);colnames(sh)=paste0(colnames(sh),"_repeat")
  rmskinfo=cbind(qh,sh)
  rmskinfo=rmskinfo[,c('ID','repClass_repeat')];colnames(rmskinfo)[colnames(rmskinfo)%in%'repClass_repeat']= ColumnName      
  
  #merge rmskinfo with peakstable
  peaksTable=merge(peaksTable[,!colnames(peaksTable)%in%ColumnName],rmskinfo,by='ID',all.x=T)
  
  #check dups and unique peaks
  dup=unique(peaksTable[duplicated(peaksTable$ID),'ID'])
  peaksTable_single=peaksTable[!(peaksTable$ID%in%dup),]
  
  peaksTable_double=peaksTable[(peaksTable$ID%in%dup),]
  u=unique(peaksTable_double$ID)
  
  #collapse duplicate peaks
  peaksTable_colapsed=as.data.frame(matrix(nrow=length(u),ncol=ncol(peaksTable_double)))
  colnames(peaksTable_colapsed)=colnames(peaksTable_double)
  
  #for each unique peak
  
  print(paste0('Annotated Regions: ',nrow(u)))
  for(x in 1:length(u)){
    peaksTable = u[x]
    pam = peaksTable_double[peaksTable_double$ID%in%peaksTable,]
    
    #if the cols of double are not in ColumnName list
    peaksTable_colapsed[x,colnames(peaksTable_double)%in%c(ColumnName)==F] = (pam[1,colnames(peaksTable_double)%in%c(ColumnName)==F,drop=T])
    
    #past the uqique values of the col to collapsed table
    peaksTable_colapsed[x,ColumnName]=paste(sort(unique((pam[,ColumnName]))), collapse =" | ")
  }
  
  #bind collapsed and unqiue
  peaksTable_colapsed = rbind(peaksTable_colapsed,peaksTable_single)
  
  #if there is no ID remove
  peaksTable_colapsed = peaksTable_colapsed[!is.na(peaksTable_colapsed$ID),]
  
  ###phil why are we adding NA columns ## This is a check to make sure rows with empty annotations get an NA
  peaksTable_colapsed[is.na(peaksTable_colapsed[,ColumnName]),ColumnName]=NA
  peaksTable_colapsed[peaksTable_colapsed[,ColumnName]%in%"",ColumnName]=NA
  
  
  return(peaksTable_colapsed) 
}

rpmsk_calling <- function(peaks_in,nmeprfix){
  ###fix variable names
  if(length(args)==0){
    nmeprfix="Same_"
    peaks_in=intronexon_Same
  }
  
  PeaksdataOut=peaks_in[!is.na(peaks_in$ID),]
  
  rpmsk_list = c("LTR","DNA","Satellite","Simple_repeat","Low_complexity","Other","Unknown")
  rpmsk_names = c("Repeat_LTR","Repeat_DNA","Repeat_Satalites","Repeat_Simple_Repeats",
                  "Repeat_Low_Complexity","Repeat_Other","Repeat_Unknown")
  i = 1
  for(find_type in rpmsk_list){
    rmsk_subset = base::subset(rmsk_GRCm38, repClass == find_type)
    PeaksdataOut=rpmsk_anno(paste0(nmeprfix,rpmsk_names[i]),rmsk_subset,PeaksdataOut)
    i = i+1
  }
  
  #run with final list
  PeaksdataOut=rpmsk_anno(paste0(nmeprfix,'Repeat_LINE_SINE'),
                          rmsk_GRCm38[rmsk_GRCm38$repClass%in%c('LINE','SINE'),],
                          PeaksdataOut)
  
  ###fix variable naming  
  p=PeaksdataOut
  p[,paste0(nmeprfix,'Repeat_comb')]=NA
  repcol=paste0(nmeprfix,c('Repeat_LINE_SINE','Repeat_LTR','Repeat_DNA',
                           'Repeat_Satalites','Repeat_Simple_Repeats','Repeat_Low_Complexity',
                           'Repeat_Other','Repeat_Unknown'))
  
  dup=p[rowSums((p[,repcol]>0),na.rm = T)>0,'ID']
  
  #run duplicated and unique
  peaksTable_single=p[!(p$ID%in%dup),]
  peaksTable_double=p[(p$ID%in%dup),]
  
  u=unique(peaksTable_double$ID)
  peaksTable_colapsed=as.data.frame(matrix(nrow=length(u),ncol=ncol(peaksTable_double)));
  colnames(peaksTable_colapsed)=colnames(peaksTable_double)
  
  #for each unique id
  for(x in 1:length(u)){
    p=u[x]
    pam=peaksTable_double[peaksTable_double$ID%in%p,]
    pam_c=((pam[1,repcol,drop=F]))
    
    pmat=as.data.frame(matrix(nrow=length(pam_c),ncol=1))
    
    ###fix col name 
    colnames(pmat)='a'
    pmat$a=t(pam_c)
    
    #remove na's
    pmat=pmat[is.na(pmat)==F,]
    
    #paste uniques
    pmat=paste(unique(pmat),collapse = ',')
    
    ###phil what is this doing?
    peaksTable_colapsed[x,colnames(peaksTable_double)]=(pam[1,colnames(peaksTable_double),drop=T])
    peaksTable_colapsed[x,paste0(nmeprfix,'Repeat_comb')]=pmat
    
  }
  
  #bind collapsed with single
  peaksTable_colapsed=rbind(peaksTable_colapsed,peaksTable_single)
  
  #remove na's
  peaksTable_colapsed=peaksTable_colapsed[!is.na(peaksTable_colapsed$ID),]
  
  ###fix variable naming
  p=peaksTable_colapsed
  PeaksdataOut=peaksTable_colapsed
  
  return(PeaksdataOut)
}

##########################################################################################
############### Assigning Clip peak attributes by strand
##########################################################################################
# Not all Peaks overlap with a single feature so peak assignments were assigned by priority:
# ncRNA > Protein coding : Exonic > repeats > Pseudogene > Antisense Feature > 
#Protein Coding : Intronic > lncRNA > no Feature  
# All annotations from RNA type, Repeat regions, and Intronic/exonic regions are annoted in the Table
peak_attributes <- function(peaks_in,nmeprfix){
  if(length(args)==0){
    nmeprfix="Same_"
    peaks_in=rpmsk_Same
    # colnames(peaks_in)
  }  
  #convert lincRNA and lncRNA types
  col_list = c("type_simple_comb","type_comb","gene_type","gene_type_ALL")
  type_start = c("lincRNA","lncRNA")
  type_end = c("linLcRNA","lnLcRNA")
  
  #for each col selected -
  for (cols_to_change in col_list){
    
    #change the type (type_list) to new id (change_list)
    for (i in range(1:length(type_start))){
      peaks_in[,paste0(nmeprfix,cols_to_change)] = gsub(type_start[i],
                                                        type_end[i],
                                                        peaks_in[,paste0(nmeprfix,cols_to_change)] )
    }
  }
  
  AddTypetoComb <- function(df_in,subset_df,type_id){
    df_in[subset_df,paste0(nmeprfix,'Comb_type_exon')]=type_id
    return(df_in)
  }
  
  #create column
  peaks_in[,paste0(nmeprfix,'Comb_type_exon')]="TBD"
  
  # 1. ncRNA
  peaks_in = AddTypetoComb(peaks_in,
                           grepl('ncRNA',peaks_in[,paste0(nmeprfix,'type_simple_comb')]),
                           "ncRNA")
  
  # 2. protein coding - Exonic
  peaks_in = AddTypetoComb(peaks_in,
                           peaks_in[,paste0(nmeprfix,'Comb_type_exon')] == "TBD" & 
                             grepl('protein_coding',peaks_in[,paste0(nmeprfix,'type_simple_comb')]) & 
                             grepl('exon',peaks_in[,paste0(nmeprfix,'feature')]),
                           "protein_coding: exon")
  
  # 3. repeats
  ###phil check the repeat_comb is correct
  peaks_in = AddTypetoComb(peaks_in,
                           peaks_in[,paste0(nmeprfix,'Comb_type_exon')] == "TBD" & 
                             is.na(peaks_in[,paste0(nmeprfix,'Repeat_comb')])==F,
                           "Repeat Element")
  
  # 4. Pseudogene
  peaks_in = AddTypetoComb(peaks_in,
                           peaks_in[,paste0(nmeprfix,'Comb_type_exon')] == "TBD" &
                             grepl('pseudogene',peaks_in[,paste0(nmeprfix,'type_simple_comb')]),
                           "pseudogene")
  
  # 5. lncRNA-exon
  peaks_in = AddTypetoComb(peaks_in,
                           peaks_in[,paste0(nmeprfix,'Comb_type_exon')] == "TBD" &
                             (grepl('linLcRNA-exon',peaks_in[,paste0(nmeprfix,'type_comb')]) |
                                grepl('lnLcRNA',peaks_in[,paste0(nmeprfix,'type_comb')]) ),
                           "lnLcRNA-exon")
  
  # 6. intron
  peaks_in = AddTypetoComb(peaks_in,
                           peaks_in[,paste0(nmeprfix,'Comb_type_exon')] == "TBD" &
                             grepl('protein_coding',peaks_in[,paste0(nmeprfix,'type_simple_comb')]),
                           "protein_coding: Intron")
  
  # 7. lncRNA-intron
  peaks_in = AddTypetoComb(peaks_in,
                           peaks_in[,paste0(nmeprfix,'Comb_type_exon')] == "TBD" &
                             grepl('linLcRNA-intron',peaks_in[,paste0(nmeprfix,'type_comb')]),
                           "lnLcRNA-intron")
  
  #all other features
  peaks_in = AddTypetoComb(peaks_in,
                           peaks_in[,paste0(nmeprfix,'Comb_type_exon')] == "TBD",
                           "no Feature")
  
  #create factor of col
  peaks_in[,paste0(nmeprfix,'Comb_type_exon')] = 
    factor(peaks_in[,paste0(nmeprfix,'Comb_type_exon')], 
           levels = c("ncRNA", "protein_coding: exon", "Repeat Element","pseudogene",
                      "lnLcRNA-exon","Antisense Feature","protein_coding: Intron","lnLcRNA-intron","no Feature"))
  
  #create col
  peaks_in[,paste0(nmeprfix,'Comb_type_ncRNA')]=NA
  
  #select ncRNA
  p_ncrna=  p_ncrna=peaks_in[peaks_in[,paste0(nmeprfix,'Comb_type_exon')]=='ncRNA',]
  
  #make col type_comb with prefix  
  p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=p_ncrna[,paste0(nmeprfix,'type_comb')]
  
  #remove types
  ### Protein coding + ncRNA annotations -> ncRNA subtype
  ### any ncRNA takes priority over pseudogene
  ### any double annotations with lncRNA become second annotation only
  type_list = c("protein_coding,",",protein_coding","pseudogene,",",pseudogene",'lnLcRNA,')
  for (type_id in type_list){
    p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub(type_id,"",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  }  
  
  #merge linLcRNA  
  p_ncrna[p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]%in%'linLcRNA-intron,linLcRNA-exon',paste0(nmeprfix,'Comb_type_ncRNA')]="linLcRNA"
  p_ncrna[p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]%in%'linLcRNA-exon,linLcRNA-intron',paste0(nmeprfix,'Comb_type_ncRNA')]="linLcRNA"
  
  type_list = c("lnLcRNA,",'linLcRNA,','linLcRNA-exon,',"linLcRNA-intron,")
  for (type_id in type_list){
    p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub(type_id,"",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  } 
  
  
  #### This section will give priority to ncRNA assignment based on order listed in ncRNAPriority
  #### This replaces previous approach to replace on a case by case baseis
  ncRNAPriority=c(
    'yRNA',
    'snRNA',
    'snoRNA',
    'srpRNA',
    'tRNA',
    '7SKRNA',
    'scRNA',
    'sncRNA',
    'scaRNA',
    'miRNA',
    'vaultRNA',
    'miscRNA',
    'misc_RNA',
    'rRNA',
    'linLcRNA-intron',
    'linLcRNA-exon',
    'lnLcRNA') 
  
  for (x in 1:length(ncRNAPriority)) {
    nc=paste0(",",ncRNAPriority[x],"|",ncRNAPriority[x],",")
    p_ncrna[grep(nc,p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]),paste0(nmeprfix,'Comb_type_ncRNA')]=ncRNAPriority[x]
  }
  
  p_ncrna[grep('miscRNA',p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]),paste0(nmeprfix,'Comb_type_ncRNA')]='misc_RNA'
  
  
  # ### any double annotations with yRNA become yRNA only
  # p_ncrna[grep('yRNA',p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]),paste0(nmeprfix,'Comb_type_ncRNA')]='yRNA'
  # 
  # ### tRNA takes priority over miRNA
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('miRNA,tRNA',"miRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # 
  # ### miRNA takes priority over rRNA
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('miRNA,rRNA',"miRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # 
  # ### snRNA takes priority over 7SKRNA
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('7SKRNA,snRNA',"snRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('snRNA,7SKRNA',"snRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # 
  # ### 7SKRNA takes priority over lncRNA
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('7SKRNA,linLcRNA-intron',"7SKRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub(',linLcRNA-intron,7SKRNA',"7SKRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('7SKRNA,linLcRNA-exon',"7SKRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('linLcRNA-exon,7SKRNA',"7SKRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('7SKRNA,lnLcRNA',"7SKRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('lnLcRNA,7SKRNA',"7SKRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # 
  # ### tRNA takes priority over rRNA
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('rRNA,tRNA',"tRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('tRNA,rRNA,',"tRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # 
  # ### snoRNA takes priority over miRNA
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('miRNA,snoRNA',"snoRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('snoRNA,miRNA,',"snoRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # 
  # ### scaRNA takes priority over miRNA
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('miRNA,scaRNA',"scaRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  # p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')]=gsub('scaRNA,miRNA,',"scaRNA",p_ncrna[,paste0(nmeprfix,'Comb_type_ncRNA')])
  
  #select not ncRNA
  p2=peaks_in[!peaks_in[,paste0(nmeprfix,'Comb_type_exon')]%in%'ncRNA',]
  
  ###fix variable
  PeaksdataOut=rbind(p_ncrna,p2)
  
  return(PeaksdataOut)
}

##########################################################################################
############### Assigning Clip peak attributes - merged strands
########################################################################################## 
Assign_CLIP_attributes=function(PeaksdataOut){
  
  #create column
  PeaksdataOut$Comb_type_exon_Oppo="no Feature"
  
  #select subset of df, input specific type
  AddTypetoComb <- function(df_in,subset_df,type_id){
    df_in[subset_df,paste0('Comb_type_exon_Oppo')]=type_id
    return(df_in)
  }
  
  # 1. ncRNA        
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               grepl('ncRNA',PeaksdataOut[,paste0('Same_','type_simple_comb')],fixed = T), 
                               'ncRNA')
  
  # 2. protein coding - Exonic
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 grepl('protein_coding',PeaksdataOut[,paste0('Same_','type_simple_comb')]) & 
                                 grepl('exon',PeaksdataOut[,paste0('Same_','feature')]),
                               'protein_coding: exon')
  
  # 3. repeats
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" &
                                 (is.na(PeaksdataOut[,paste0('Same_','Repeat_comb')])==F),
                               "Repeat Element")
  
  # 4. Pseudogene
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 grepl('pseudogene',PeaksdataOut[,paste0('Same_','type_simple_comb')]),
                               "pseudogene")
  
  # 5. Antisense - non LncRNA
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 ((PeaksdataOut[,paste0('Oppo_','Comb_type_exon')]%in%'no Feature')==F) &
                                 grepl('lnLcRNA',PeaksdataOut[,paste0('Oppo_','Comb_type_exon')])==F,
                               "Antisense Feature")
  
  # 6. lncRNA-exon
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 (grepl('linLcRNA-exon',PeaksdataOut[,paste0('Same_','type_comb')]) | 
                                    grepl('lnLcRNA',PeaksdataOut[,paste0('Same_','type_comb')]) ),
                               'lnLcRNA-exon')
  # 7. Antisense - LncRNA - exon
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 ((PeaksdataOut[,paste0('Oppo_','Comb_type_exon')]%in%'no Feature')==F & 
                                    grepl('intron',PeaksdataOut[,paste0('Oppo_','Comb_type_exon')])==F ),
                               'Antisense Feature')
  
  # 8. intron
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 grepl('protein_coding',PeaksdataOut[,paste0('Same_','type_simple_comb')]),
                               'protein_coding: Intron')
  
  # 9. Antisense - LncRNA - intron
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 ((PeaksdataOut[,paste0('Oppo_','Comb_type_exon')]%in%'no Feature')==F) & 
                                 (grepl('linLcRNA-exon',PeaksdataOut[,paste0('Oppo_','Comb_type_exon')])==F),
                               'Antisense Feature')
  
  
  
  # 10. lncRNA -intron
  PeaksdataOut = AddTypetoComb(PeaksdataOut, 
                               PeaksdataOut[,'Comb_type_exon_Oppo'] == "no Feature" & 
                                 grepl('linLcRNA-intron',PeaksdataOut[,paste0('Same_','type_comb')]),
                               'lnLcRNA-intron')
  
  PeaksdataOut[,paste0(c('Same_'),'Comb_type_exon')]= gsub('lnLcRNA-exon','lncRNA',PeaksdataOut[,paste0(c('Same_'),'Comb_type_exon')] )
  PeaksdataOut[,paste0(c('Oppo_'),'Comb_type_exon')]= gsub('lnLcRNA-exon','lncRNA',PeaksdataOut[,paste0(c('Oppo_'),'Comb_type_exon')] )
  PeaksdataOut[,paste0(c('Same_'),'Comb_type_exon')]= gsub('lnLcRNA-intron','lncRNA',PeaksdataOut[,paste0(c('Same_'),'Comb_type_exon')] )
  PeaksdataOut[,paste0(c('Oppo_'),'Comb_type_exon')]= gsub('lnLcRNA-intron','lncRNA',PeaksdataOut[,paste0(c('Oppo_'),'Comb_type_exon')] )
  PeaksdataOut[,'Comb_type_exon_Oppo']= gsub('lnLcRNA-exon','lncRNA',PeaksdataOut[,'Comb_type_exon_Oppo'] )
  PeaksdataOut[,'Comb_type_exon_Oppo']= gsub('lnLcRNA-intron','lncRNA',PeaksdataOut[,'Comb_type_exon_Oppo'] )
  PeaksdataOut[,'Comb_type_exon_Oppo']=factor(PeaksdataOut[,'Comb_type_exon_Oppo'], levels = c("ncRNA", "protein_coding: exon", "Repeat Element","pseudogene","lncRNA-exon","Antisense Feature","protein_coding: Intron","lncRNA-intron","lncRNA","no Feature"))
  
  return(PeaksdataOut)
}

##########################################################################################
############### Add annotations to junctions
##########################################################################################  
# Join_Junction_Combine=function(Peaksdata2_anno_in,read_depth,FtrCount_out,anchor) {
#   # Peaksdata2_anno_in=Peaksdata2_anno
#   # anchor=anno_anchor
#   
#   Junc_peaks=unique(FtrCount_out$Junc_peaks)
#   PGene_TBL2=FtrCount_out$PGene_TBL2
  
Join_Junction_Combine=function(Peaksdata2_anno_in,read_depth,Junc_peaks,PGene_TBL2,anchor) { 
    # Peaksdata2_anno_in=Peaksdata2_anno
    # anchor=anno_anchor
  Junc_peaks=Junc_peaks$x
  
  Peaksdata2_anno_trns_exon=( Peaksdata2_anno_in[Peaksdata2_anno_in$ID%in%Junc_peaks,])
  CollapsedOut=as.data.frame(matrix(nrow = (1),ncol = ncol(Peaksdata2_anno_trns_exon)))
  colnames(CollapsedOut)=colnames(Peaksdata2_anno_trns_exon)
  CollapsedOut$IDmerge=NA
  
  for (x in 1:nrow(PGene_TBL2)) {
    Trnsc=PGene_TBL2[x,]
    TrnscUL=unlist(unique(str_split(Trnsc$JoinID,pattern = ",")))
    
    d2=Peaksdata2_anno_trns_exon[Peaksdata2_anno_trns_exon$ID%in%TrnscUL,]
    
    if (nrow(d2)==0) {
      next
    } else if (unique(d2$strand)=="+") {
      d5=d2[d2$start%in%min(d2$start),] 
      dmax=d2[d2$Counts_fracMM%in%max(d2$Counts_fracMM),]
      if (nrow(dmax)>1) {dmax=dmax[dmax$start%in%min(dmax$start),]}
      dmaxT=d2[d2$Counts_total%in%max(d2$Counts_total),]
      if (nrow(dmaxT)>1) {dmaxT=dmaxT[dmaxT$start%in%min(dmaxT$start),]}
    } else if (unique(d2$strand)=="-") {
      d5=d2[d2$start%in%max(d2$start),]
      dmax=d2[d2$Counts_fracMM%in%max(d2$Counts_fracMM),]
      if (nrow(dmax)>1) {dmax=dmax[dmax$start%in%max(dmax$start),]}
      dmaxT=d2[d2$Counts_total%in%max(d2$Counts_total),]
      if (nrow(dmaxT)>1) {dmaxT=dmaxT[dmaxT$start%in%max(dmaxT$start),]}
    }
    
    ### collapse all columns 
    collapseCol=function(x){
      # x=d2[,'Same_feature',drop=F]
      if (T%in%grepl(",",x)) {
       x=x[!is.na(x)]
       x= str_split(x,",")%>%unlist%>%unique%>%sort
       x=paste0(x,collapse=",")
      }else{
        x=x[!is.na(x)]
        x= paste0(unique(x),collapse=",")
      }
      return(x)  
    }
    
    # d3= ( apply(d2 ,2, function(x){paste(unique(x[!is.na(x)]),collapse = ',')}))
    d3= apply(d2 ,2, collapseCol)
  
    ### add values from all regions
    d3[d3==""]<-NA
    cols=c('Counts_Unique','Counts_fracMM','Length')
      ## convert to numeric
    d3=as.data.frame(t(d3))
    d3[1,cols]=t(as.data.frame(colSums(d2[,cols])))   
    d3[,cols]=as.numeric(d3[,cols])
    d3$IDmerge=d3$ID
    
    ## Select summary annotation based on Max(dmax) expression peak or 5'(d5) peak   
    if (anchor=="max_MM") {
      danno=dmax
    }else if (anchor=="max_total"){
      danno=dmaxT
    }else if (anchor=="5prime"){
      danno=d5
    }else { print("select either max or 5prime for AnnoAnchor paramater")}
    
    cols_AnnotSelect=c('Same_Comb_type_exon','Same_Comb_type_ncRNA','Oppo_Comb_type_exon',
                       'Oppo_Comb_type_ncRNA','Comb_type_exon_Oppo')
    d3[,cols_AnnotSelect]= (danno[,cols_AnnotSelect])
    
    # Select ID for 5' most read or max counts
    d3$start=danno$start
    d3$end=danno$end
    d3$ID=paste0(danno$chr,":",danno$start,"-",danno$end,"_",danno$strand)   
    
    ## if count < read_depth skip
    if (d3$Counts_fracMM<read_depth) {
      next
    } else{
      CollapsedOut=rbind(CollapsedOut[,colnames(CollapsedOut)],d3[1,colnames(CollapsedOut)])
    }
  }  
  CollapsedOut=CollapsedOut[is.na(CollapsedOut$ID)==F,]
  Peaksdata2_anno_in$IDmerge=NA
  Peaksdata2_anno_in=Peaksdata2_anno_in[Peaksdata2_anno_in$ID%in%Junc_peaks==F,]
  Peaksdata2_anno_in=(rbind(Peaksdata2_anno_in,CollapsedOut))
  
  
  ######################################################################
  #### Re filter reads after combing counts of spliced reads
  Peaksdata2_anno_in=Peaksdata2_anno_in[Peaksdata2_anno_in$Counts_fracMM>=read_depth,]
  
  return(Peaksdata2_anno_in)
}    

#### collapse exon
Collapse_Exon=function(Peaksdata2_anno){    
  Peaksdata2_anno_trns=Peaksdata2_anno[(Peaksdata2_anno$Same_feature)%in%""==F,]
  Peaksdata2_anno_trns_exon=Peaksdata2_anno_trns[Peaksdata2_anno_trns$Comb_type_exon_Oppo%in%'protein_coding: exon',]
  trns=Peaksdata2_anno_trns_exon$Same_ensembl_gene_id
  trns=trns[grep(",",trns)]
  trns=c(Peaksdata2_anno_trns_exon$Same_ensembl_gene_id,unlist(strsplit(trns,",")))
  dupGeneName=unique(trns[duplicated(trns)])
  
  if(length(dupGeneName)>0){
    CollapsedOut=as.data.frame(matrix(nrow = 1,ncol=ncol(Peaksdata2_anno_trns_exon)));colnames(CollapsedOut)=c(colnames(Peaksdata2_anno_trns_exon))
    for (x in 1:length(dupGeneName)) {
      d1=dupGeneName[x]
      d2=Peaksdata2_anno_trns_exon[grep(d1,Peaksdata2_anno_trns_exon$Same_ensembl_gene_id),]
      
      if (nrow(d2)==0) {
        next
      } else {
        d2$ID=paste0(d2$chr[1],":",min(d2$start),"-",max(d2$end),"_",unique(d2$strand))
        
        if (unique(d2$strand)=="+") {
          d3=d2[d2$start%in%min(d2$start),]
        } else if (unique(d2$strand)=="-") {
          d3=d2[d2$start%in%max(d2$start),]
        }
        
        d3[,'start']=min(d2$start)
        d3[,'end']=max(d2$end)
        d3$IDmerge=paste(d2$ID,collapse = ",")
        
        trns=trns[grep(",",d2$Same_ensembl_gene_id)]
        d3$Same_ensembl_gene_id=paste0(unique(c(d2$Same_ensembl_gene_id,unlist(strsplit(trns,",")))),collapse = ",")
        trns=trns[grep(",",d2$Same_gene_name_comb)]
        d3$Same_gene_name_comb=paste0(unique(c(d2$Same_gene_name_comb,unlist(strsplit(trns,",")))),collapse = ",")
        trns=trns[grep(",",d2$`Opposite Strand: Host_gene_ensembl_id`)]
        d3$Oppo_ensembl_gene_id=paste0(unique(c(d2$Oppo_ensembl_gene_id,unlist(strsplit(trns,",")))),collapse = ",")
        trns=trns[grep(",",d2$Oppo_gene_name_comb)]
        d3$Oppo_gene_name_comb=paste0(unique(c(d2$Oppo_gene_name_comb,unlist(strsplit(trns,",")))),collapse = ",")
        
        d3$Same_ensembl_gene_id=gsub('NA,|,NA',"",d3$Same_ensembl_gene_id)
        d3$Same_gene_name_comb=gsub('NA,|,NA',"",d3$Same_gene_name_comb)
        d3$Oppo_ensembl_gene_id=gsub('NA,|,NA',"",d3$Oppo_ensembl_gene_id)
        d3$Oppo_gene_name_comb=gsub('NA,|,NA',"",d3$Oppo_gene_name_comb)
        
        CollapsedOut=rbind(CollapsedOut,d3)
      }
    }
    
    rownames(CollapsedOut)=NULL
    CollapsedOut=CollapsedOut[-1,]
    CollapsedOutID= unique(unlist(strsplit(CollapsedOut$IDmerge,",")))
    Peaksdata2_anno$IDmerge=NA
    Peaksdata2_anno=Peaksdata2_anno[Peaksdata2_anno$ID%in%CollapsedOutID==F,]
    Peaksdata2_anno=(rbind(Peaksdata2_anno,CollapsedOut))
    
  } else { 
    print("Multiple peaks not found in single Gene")
  }
  return(Peaksdata2_anno)
} 

##########################################################################################
## correct lnLc annotation - keep lnLc annotation above to hepl track source of lnc annotation
##########################################################################################
correct_lnLc=function(Peaksdata2_anno){
  for (x in c('Same_gene_type','Same_gene_type_ALL','Same_type_simple_comb','Same_type_comb',
              'Oppo_gene_type','Oppo_gene_type_ALL','Same_type_simple_comb','Oppo_type_simple_comb','Oppo_type_comb')){
    
    Peaksdata2_anno[,x]=gsub("lnLcRNA","lncRNA",Peaksdata2_anno[,x])
    Peaksdata2_anno[,x]=gsub("linLcRNA","lincRNA",Peaksdata2_anno[,x])
  }
  return(Peaksdata2_anno)
}
